pyproject.toml
[tool.poetry]
name = "herbarium-repository-databot"
version = "0.1.0"
description = ""
authors = [
    "Petr Novotný <krkabol@gmail.com>"
]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2"
pyyaml = "^6"
flask = "^3"
APScheduler = "^3"